// src/app/interfaces/company-registration.ts
export interface CompanyRegistration {
    username: string | null | undefined ; 
    password: string | null | undefined ; 
    email: string | null | undefined ; 
    company_label: string | null | undefined ; 
    user_phone_number: string | null | undefined ; 
    user_join_date: string | null | undefined ; 
    creation_date: string | null | undefined;
    notifications:  null;
    alerts: null;
    companyProfile:  null;
}